added icons for custom admonitions
