+++ website/src/css/custom-admonition.css
@@ -25,8 +25,16 @@
 }
 
 .admonition-icon {
-  margin-right: 0.5rem;
   display: inline-block;
+  vertical-align: middle;
+  margin-right: 0.4em;
+}
+
+.admonition-icon svg {
+  display: inline-block;
+  height: 1.6em;
+  width: 1.6em;
+  fill: var(--ifm-alert-foreground-color);
 }
 
 /* Custom admonition types: function and class */
